Scale track capacity when required

+++ arctracker-engine/src/loader/format_desktop_tracker.c
@@ -175,11 +175,11 @@ static bool decode_dtt_patterns(uint8_t *base_address, const uint32_t *pattern_o
         uint8_t *raw_pattern_data = base_address + pattern_offsets[pno];
         for (int line = 0; line < pattern_length; line++)
         {
-            for (int track = 0; track < module->track_capacity; track++)
+            for (uint32_t track = 0; track < module->track_capacity; track++)
             {
-                const int event_index = (line * module->track_capacity) + track;
+                const uint32_t event_index = (line * module->track_capacity) + track;
                 event_t *event = module->patterns[pno].events + event_index;
-                if (track < module->num_tracks)
+                if (track < (uint32_t) module->num_tracks)
                     raw_pattern_data += decode_desktop_tracker_event(raw_pattern_data, event);
                 else
                     *event = (event_t) {0};
